requirements/base.txt
Django==2.1.2
django-cors-headers==2.4.0
django-filter==2.0.0
djangorestframework==3.9.0
graphene==2.1.3
graphene-django==2.2.0
mysqlclient==1.3.12
